Added Important Info on Bundle and Grouped Products
The source deduction service use the deprecated Catalog Inventory StockItem model, which holds the In Stock product attribute. However, this requires the simple products of those bundle and groups, to be also assigned to the default stock, and they all must be marked as In Stock at that source. 

For reference, the issue lies here: vendor/magento/module-inventory-configuration/Model/GetLegacyStockItem.php