Store Fulfillment Blue Acorn changes (#282)
* Added Important Info on Bundle and Grouped Products

The source deduction service use the deprecated Catalog Inventory StockItem model, which holds the In Stock product attribute. However, this requires the simple products of those bundle and groups, to be also assigned to the default stock, and they all must be marked as In Stock at that source. 

For reference, the issue lies here: vendor/magento/module-inventory-configuration/Model/GetLegacyStockItem.php

* Walmart no suuport for 2FA

The 2FA support for Adobe Commerce exists but Walmart is not supporting the service and is out of scope for the time being affects all versions 1.0.0 - 1.30

* Important integration token database restore

Merchants need to have a best practice approach if restoring environment data using another environments data source. Overwriting integration will require merchant to be re-onboarded with Walmart systems.

* Requires Bearer token for OAuth

Requires Bearer token for OAuth in order to make a successfull call over API

* Update help/store-fulfillment/inventory-stock-transfer.md

* Update help/store-fulfillment/connect-set-up-service.md

* Editorial fixes

* Fix markdown syntax

* Update help/store-fulfillment/connect-set-up-service.md

Co-authored-by: Shrie Spangler <spangler@adobe.com>

Co-authored-by: Luis Magaña <luisuxcli@gmail.com>
Co-authored-by: Shrie Spangler <spangler@adobe.com>