Update events.md
Some suggested edits in the profile events section for consistency vs. internal wiki page, and for accuracy to the XDM schema. Note on person.personalEmailID - the XDM description is the same as for personalEmail.address, but I do think it might be confusing to have 2 fields listed with the same description here. Will leave that to your judgement.